backend/backend/requirements.txt
Django>=5.2.5,<6
gunicorn
whitenoise
psycopg2-binary
django-cors-headers
dj-database-url
python-dotenv